Fix full & summarize to actually work.
This was a bit weird because simply printing the map didn't show all the
keys but converting it to another datastructure (e.g. via `seq` as
cheshire does internally) did (more or less obviously).

One downside is that after using `full` or `summarize` on can't get
additional attributes, e.g. we can't make our queries return "sparse"
entities by default and extend only on request. But maybe that's a good
thing because we don't taint our queries with such concerns... We'll se.